files: Move files viewer bits to a separate widget
Issue #290
